frontend

+++ backend/controllers/noteController.js
@@ -2,7 +2,7 @@ const asyncHandler = require('express-async-handler');
 const Note = require('../models/noteModel');
 
 const getNotes = asyncHandler(async (req, res) => {
-    const notes = await Note.find({});
+    const notes = await Note.find({user: req.user.id});
     res.json(notes);
     res.status(200).send();
 })
@@ -12,7 +12,10 @@ const setNote = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Text is required')
     }
-    const note = await Note.create(req.body);
+    const note = await Note.create({
+        ...req.body,
+        user: req.user.id,
+    });
     res.status(200).json(note);
 })
 
@@ -21,6 +24,10 @@ const updateNote = asyncHandler(async (req, res) => {
     if(!note){
         res.status(404)
         throw new Error('Note not found')
+    }
+    if(note.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error('Not authorized')
     }
     const updatedNote = await Note.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
@@ -33,6 +40,10 @@ const deleteNote = asyncHandler(async (req, res) => {
     if(!note){
         res.status(404)
         throw new Error('Note not found')
+    }
+    if(note.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error('Not authorized')
     }
     const deletedNote = await Note.findByIdAndDelete(req.params.id);
     res.status(200).json(deletedNote);
